linephone-simple: port of the "SIP Client for Ubuntu Touch"
This project contains a port of 'SIP Client for Ubuntu Touch' based on
the git commit (4c03a020) [1], the 'AboutPage.qml' was adapted by hand
from the 'about.qml.in' template.

There are a variety of changes to adjust the GUI for the current
use-case as SIP show-case on the PinePhone.

[1] https://gitlab.com/ubports-linphone/linphone-simple/-/archive/4c03a020/linphone-simple-4c03a020.tar.gz